待测试订阅更新， TODO其他任务

+++ src/main/java/io/github/yajuhua/podcast2/task/TaskPackage.java
@@ -15,6 +15,6 @@ import java.util.concurrent.TimeUnit;
 public class TaskPackage {
     private Runnable task;
     private TimeUnit timeUnit;
-    private Integer timeout;
+    private long timeout;
     private String description;
 }
